+++ node/resolvers/payment.ts
@@ -1,8 +1,6 @@
 export const queries = {
-  getCardSessionId: async (_: {}, __: {}, ctx: Context): Promise<String> => {
-    const {
-      clients,
-    } = ctx
+  getCardSessionId: async (_: {}, __: {}, ctx: Context): Promise<string> => {
+    const { clients } = ctx
     const { id: cardSessionId } = await clients.checkout.getPaymentSession()
     return cardSessionId
   },
@@ -11,12 +9,14 @@ export const queries = {
 export const mutations = {
   savePaymentToken: async (
     _: {},
-    { paymentTokens } : { paymentTokens: { creditCardToken: string, paymentSystem: string }[] },
+    {
+      paymentTokens,
+    }: {
+      paymentTokens: Array<{ creditCardToken: string; paymentSystem: string }>
+    },
     ctx: Context
   ): Promise<SavePaymentTokenPayload> => {
-    const {
-      clients
-    } = ctx
+    const { clients } = ctx
     const { checkout } = clients
     await checkout.savePaymentToken(paymentTokens)
     return { status: 'OK' }