+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update to node builder 6.x.
+- Call graphql gateway instead of search-graphql directly.
 
 ## [0.33.1] - 2020-04-27
 ### Fixed